+++ R/components.R
@@ -1185,7 +1185,10 @@ component_proximity_selected <- function(
         ) %>%
         ggplot(aes(x = sample_alias, y = !!sym(proximity_score))) +
         geom_hline(yintercept = 0) +
-        geom_violin(aes(fill = condition), draw_quantiles = 0.5, drop = FALSE) +
+        geom_violin(aes(fill = condition),
+                    draw_quantiles = 0.5,
+                    drop = FALSE,
+                    scale = "width") +
         geom_jitter(
           size = 0.1,
           position = position_jitter(height = 0),